requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0

# Redis
redis==5.0.1
rq==1.15.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2

# Document Processing
pypdf==3.17.0
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10
langdetect==1.0.9
spacy==3.7.2

